make: add perf-deb-src-pkg target to build perf debian source package
Alternative approach to the previous attempt. Squashing perf
build (which is userspace tool) to kernel deb-pkg target makes things
just ugly and fragile. The simple handcrafted dpkg-deb calls work fine
for kernel, but perf has shared library dependencies etc.

Generate the perf source package independent of the kernel package, and use
regular debhelper tools to make a nice userspace package. The end result
is a a short script of almost nothing but shell Here documents.

The beginning of the script is copypaste of parts of builddeb script.
Proper submission would include sharing these bits. But for the RFC
the new script is self-contained to be easier to understand. ...And to
be useful outside the kernel sources in case nobody else finds building
perf packages out of upstream kernel trees important.

Signed-off-by: Riku Voipio <riku.voipio@linaro.org>